+++ include/moveinfo.h
@@ -5,10 +5,10 @@
 class MoveInfo {
 public:	
 	int captured_piece;
-	uint64_t enpassant_square;
+	int enpassant_square;
 	int castling_rights;
 
-	MoveInfo(int, uint64_t, int);
+	MoveInfo(int, int, int);
 	~MoveInfo();
 
 private:
