+++ src/position.cpp
@@ -1,6 +1,6 @@
 #include "position.h"
 
-Position::Position(uint64_t* piece_list, int color, int castling_rights, uint64_t enpassant_square, int half_moves, int full_moves) {
+Position::Position(uint64_t* piece_list, int color, int castling_rights, int enpassant_square, int half_moves, int full_moves) {
 	for (int i = 0; i < 12; i++) {
 		Position::piece_list[i] = piece_list[i];
 	}
